fix(tgnms): fix doublescrollbar on NetworkNodesTable and NetworkPlanningTable
Summary:
Material-table is statically sized using the maxBodyHeight tableoption. If the networktables view is too short, like on the maps page, there will be a scrollbar for the table itself and another for the view the table is sitting in.

This dynamically sets the height by subtracting the height of the sections above the table. The NetworkTables component first subtracts the height of the Tabs before passing it along to the individual tables. The individual tables then further subtract the height to match the components they render. The height of the table changes based on whether the toolbar, filtering, or grouping is enabled.

Scrollbars galore Before:
https://pxl.cl/1GsnL

One scrollbar After:
https://pxl.cl/1GsnV
https://pxl.cl/1GsnT
https://pxl.cl/1GsnS
https://pxl.cl/1GsnX
https://pxl.cl/1GsnW

Reviewed By: cjminer505

Differential Revision: D27630486

fbshipit-source-id: 01579d199dc7f2074cdd019f4e17f9b18519c950