feat(tgnms): Planning - select folder from planeditor
Summary:
Creates SelectANPFolder component to allow user to select which folder to create a plan in.

Before D27637047, all plans were created in the tgnms folder. After D27637047, plans are created in the folder that's currently selected in the folders table. If the user clicks "Plan new Area" with no folder selected, they can't create the plan. This creates a SelectFolder component for the PlanEditor.

https://pxl.cl/1GQCZ

https://pxl.cl/1GQD0

Reviewed By: cjminer505

Differential Revision: D27795053

fbshipit-source-id: b5fe4b92924dc2453a51c0a452afdfeb338b6705