Reduce weather calls to only PoP locations
Summary: The weather service asks for weather at locations of all sites in the network, this results in too many requests for the free API. This diff reduces the no. of locations for which weather is tracked to just PoP sites. Since the nodes are geographically close to one another, this is more than sufficient for tracking weather around the network.

Reviewed By: aclave1

Differential Revision: D27924982

fbshipit-source-id: ebcd37df137d6c41a9f5cfddc66936938f4a4c6c