feat(tgnms): Annotation multiselect and bulk actions, fix some bugs
Summary:
Adds Annotation multiselect functionality to the side-panel.
* Shows an instance of the annotation editor form for each selected annotation.
* Provides bulk-delete
* Bulk convert-to-site
{F610107614}

Bulk convert-to-site
https://pxl.cl/1H4hV
## Fixes

**Annotation form not updating map:**
Whenever the user uses mapbox-draw's setFeatureProperty, the map doesnt update until the user hovers over it. The hack to get around this is to do: `drawControl.add(drawControl.get(featureId));`

Here's a video of this in action:
https://pxl.cl/1GW1F

I got the hack from: https://github.com/mapbox/mapbox-gl-draw/issues/878#issuecomment-516786006

**UI crash:**
NMS ui would crash if the user:
* enable draw mode
* select an annotation
* disable draw mode
* enable draw mode

This is because MapAnnotationContextProvider used drawControl.get() even when the draw control is disabled. That code was removed when moving to multiselect.
https://pxl.cl/1H9FT

Reviewed By: cjminer505

Differential Revision: D27835126

fbshipit-source-id: 4482add79727cfffe634177e494aebd884fe17fd