fix(tgnms): Remove ANP output file hack
Summary:
Output files were invalid json files, so I added a workaround to turn it into valid json. After T84339647 was fixed, the hack broke the UI with new output files which didn't contain the bug.

Warning: This will break visualizing of plans in the NMS map for old plans which contained this bug.

Reviewed By: cjminer505

Differential Revision: D27956339

fbshipit-source-id: ab91143049b9196a03e3cecdab26770318146e74