fix(tgnms): recompute compute node bearings failed if link deleted and readded, also fix azimuth calculation
Summary:
When a link is added, 3 api requests are made: one to create the link, and 2 to recompute the ant_azimuth of the nodes on either side of the link.

Nodes were updated with the recomputed azimuths even if the value was exactly the same, apiservice rejected the non-change. With the bug present, the link is still added but it appears to the user as though something has gone wrong. This compares the azimuths and if they're sufficiently different, it'll make the request.

Also this fixes the calculation for positive bearings.

Another issue that this still doesn't resolve is the fact that bearings are not computed when a link is deleted. This issue would've been less likely to occur (unless the user manually edits topology file or ant_azimuth) if bearings were recomputed. Created T89517150 to track that

Reviewed By: cjminer505

Differential Revision: D27976866

fbshipit-source-id: ef70004f977ab42ecd8c6edf552f4defffdab200