fix(nms): deleting link no longer disables auto ignition
Summary:
when force deleting and adding a link back, it would set ignition state to false, to turn off the link to delete, but then would not enable it again after the link was deleted.

This made it so if you deleted a link then added it back, you would have to manually enable link ignition for the link before it would turn on.

Reviewed By: elludraon, aclave1

Differential Revision: D28651359

fbshipit-source-id: 0591436f093a78e26e0b1a778ba4b63e61457702