BE(nms): Hide secrets in the reset button
Summary: Currently we display the secret in the reset button below the input field. In order to ensure security of sensitive input fields are hidden by default and shown only by the consent of the user, this diff will hide secrets that are shown in the reset button. This will allow users to safely share screen and collaborate.

Reviewed By: aclave1

Differential Revision: D29368171

fbshipit-source-id: 5f49d61d3b33cba3853743517759c32a51993da0