BE(tgnms): startup error messages
Summary:
Start handling requests ASAP in the NMS and serve a startup page until the NMS is fully initialized. Use this startup page to surface startup errors to the user instead timing out. NGINX will show a 502 error to the user when hosted in swarm/k8s.

UI:
https://pxl.cl/1KXkF

#Caveat:
One issue this feature is that the NMS will show a 200 during startup, even though it's not fully ready to serve requests. This is because in kubernetes, the pod is not considered "alive" and traffic will not be sent to it, until its liveness/readiness endpoints return a 200. Originally I had NMS returning a 503 Service Unavailable but since this caused the health-check readiness endpoint to not return a 200, no traffic would be sent to the NMS and the error page wouldn't show, instead a 502 would still show. This would make the feature useless.

Differential Revision: D29407593

fbshipit-source-id: c93df4581e359b987d087dd64a97a5ed848b013a