feat(tg-cli): pull images for docker tags during install
Summary:
This pulls the latest image for the given tag every time install is run. This does not pull :latest - This just pulls the current latest image with the given tag.

This merges the image variables from the group_vars/all file and the config.yml file.

Old Behavior:
A given tag was only pulled once. That means if we updated an image's :stable, the installer would not re-pull it. This prevented users from getting new bugfixes and added a manual step for them to script out pulling all images

Reviewed By: pmccut

Differential Revision: D29504972

fbshipit-source-id: 05d7f193380b56ad2e36d6a832364f589fbd6b72