Create show-config command and fix is_ipv6 bug
Summary:
The show-defaults (now called show-config) command used to allow users to edit the `is_ipv6` flag. This is a field that should only be modified by the installer (it is kept in the config file so that the installer knows to fetch the value for it).

The new `show-config` allows users to get the most critical config fields, but also the option to get all configurable options. It also prevents the restricted fields from showing.

Reviewed By: aclave1

Differential Revision: D29774409

fbshipit-source-id: f3d1f1baac9ec0dff17ac20eeb6d72ef1c6c68bf