BE(tgnms): add support for swarm services to develop.py
Summary:
We use this develop.py script during development to build a docker image and deploy it to a cluster. Previously it supported k8s only, this adds support for swarm by adding a new parameter to the CLI - cluster.

Pass the --cluster k8s|swarm to select which cluster you're deploying to. k8s is the default.

Usage:
* Must be run from inside the FB lab. We can change it to support other docker builders if anyone needs that.
* python ~/connectivity-lab/terragraph/nms/jenkins_scripts/develop.py go -m <cluster to deploy to> --cluster swarm --project nginx_swarm

Caveats:

* The only container that it can't automatically update is the proxy - nginx_proxy_monitor. This is currently not run through swarm, it's run through docker compose - so it's a bit harder to update the image it uses.
* This doesn't implement the "reset" command for swarm - but it should be fairly easy following this example.

Reviewed By: elludraon

Differential Revision: D30280255

fbshipit-source-id: ca6aeb8a440752c973396ba0f6dd83079f273565