fix(tgnms): send correct link data when performing topology import
Summary:
Previously, when importing a TG JSON topology, it would not import the links.

This was due to an error in the nms import code that wasn't suited for newer versions of the e2e_controller. Specifically, certain link fields were not being sent to the `bulkAdd` endpoint.

This diff makes sure the fields are being sent to the e2e_controller.

Reviewed By: aclave1

Differential Revision: D30350943

fbshipit-source-id: 0ffc4999dbbcc4af457c48b620019aba6ea977d7