Draft Plans
Summary:
Reimplement the planning API inside NMS. The NMS now stores all state and parameters for running a plan. Once a user launches the plan, all the local input files are uploaded and the plan is created and launched using the planning API.

## Why?
The planning API is immutable and stateless. Once something is created using the API, it cannot be changed. This means that a user can not edit a plan's name or input files before running it.

## What use-cases does this enable?
This enables 2 main use-cases right now, and more down the line.
1. Creating a plan and saving it for later. Previously, a user had to name a plan and upload/select all its input files and launch it in one go which is bad UX.
2. Editable input files - One of the major features we plan to support is editable input files. Users will be able to edit their sites file and boundary file. This is not possible with the current planning API and must be done inside the NMS.

# Dev notes

All database rows contain a nullable column called fbid. This is to reference the object in the planning API once it has been created.
### Database Tables:
* networkPlanFolder - Holds a group of network plans
* networkPlan - holds the parameters and references to input files for a plan. networkPlan tracks a plan through its full lifecycle, from initial creation to analysis.
* networkPlanFile - a reference to a file. A file can be either stored on the NMS's local disk (and editable) or in the planning backend (immutable)

Reviewed By: tommyhuynh

Differential Revision: D28620291

fbshipit-source-id: d3f6272633a751c160528c425fcddc941d0337cd