fix(tgnms): Down link not showing on grafana link dashboard
Summary:
Grafana populates it's variables (on the header) by making queries to prometheus and gathering unique values.

Previously: when a link would go down, it would no longer appear in that query ^ and thus it would not appear as an option in the variable.

With this change, we now are able to gather that link as an option. The user just has to make the the time range encompasses a period when the link was online.

# Technical details
We change the query to be: `query_result(snr{network=\"$network\"} offset ${__range_s}s)`
- we get all results that match `network=\"$network\"`
- we perform that query at a specific time period: the beginning of the `__range_s` (before the link went down; and thus should be in the response).
- `query_result` is a way to perform a more complex query.

We then use regex to gather the `linkName` from the response.

Reviewed By: aclave1

Differential Revision: D30679471

fbshipit-source-id: 7f1b2220e1b58cf57f25b961fb3841de25240498