feat(tgnms): Commit whole plan to topology
Summary:
# Description
Enables users to commit an entire plan to topology. Currently it will error out if there any any elements that are already in the network.

## activeANPJsonTopology

Explaining the design decision around `activeANPJsonTopology` field: this field represents the topology visible **on the map** (since users can toggle Proposed, Unavailable, Candidate).

When the user decides to commit the topology, the topology that is visible on the map is the one that is used. (The logic here is that *"what you see is what you get"*).

In the future during the multi-select stage, `activeANPJsonTopology` can hold all the elements that are selected, or something along those lines.

Reviewed By: aclave1

Differential Revision: D30581270

fbshipit-source-id: 44e0f15b2ba09ca353a084cfc1af1bfae3cf8f2a