fix(tgnms) Starting plan without saving fails
Summary:
# Bug Description
When a user creates a new plan, uploads files, then *immediately* **starts** the plan; it would fail because the plan was never saved to the DB. (Recall that launching a plan only requires the plan id and the Planning Service would then look up the details in the DB.)

If instead i created a new plan, uploaded files, **saved**, *then* **started** the plan; it would work fine.

# Fix
Save the plan when the user starts a plan.

# Other misc things in this diff
- **When a user Starts a plan, replace all the buttons with a loading sign** (since there's no turning back now anyways). Previously when clicking Start Plan, the buttons would still be clickable and it wasn't clear that something was happening in the background; it looked like the Start command didnt click).
- **Snackbar notification that a plan was successfully saved**, previously it wasn't clear whether a plan saved successfully or not. I think this change makes things feel more responsive and clear.

Reviewed By: aclave1

Differential Revision: D30704462

fbshipit-source-id: d89f5328ff523e5ed89755ad859a0b78d5b6d461