fix(tgnms): docker packages being held
Summary: Fixes an issue where installing nms would fail due to docker packages upgrading. The package was being held before it had a chance to be installed. This moves the hold command to AFTER the install is finished. This also removes the strict versioning requirements for installing the docker package as I don't believe they're necessary anymore.

Reviewed By: tommyhuynh

Differential Revision: D30809195

fbshipit-source-id: 0407e3975d110834cb57d6f430e2f7a3223026f6