Remove horizontal scroll bar from topology table in Network Planning
Summary: The topology table has a horizontal scroll bar because it has a child component with a width that is bigger than its container. Specifically, the table header, which is implemented with a Grid component, is too big. We resolve this issue by adding a width: 100% style to the grid.

Reviewed By: tommyhuynh

Differential Revision: D31181774

fbshipit-source-id: 42084e8c9dc159c99fb6cc08407e0dd72fde29dc