bug(installer): Fix uninstall not using config parameters
Summary:
**Issue:**
While uninstalling the nms, i noticed that the config file parameters were not being used in the uninstall playbook.

it would try to ssh into my VMS with user `ubuntu` when i clearly set it as `root` in my config file.

**Fix:**
We need to `generate_variables` and pass it to the runner.

I also refactored a bit so the decorator can accept an arbitrary number of options beyond the common ones.

Reviewed By: aclave1

Differential Revision: D31412564

fbshipit-source-id: f4396b29ba2605e13711215a798c4a1e9950eac1