bug(tgnms-v2) Planning: Fix ANP output file renames causing failed topology load
Summary:
# Overview
There was a bug where we were not able to load the topology for successful ANP plans because the output files were renamed, and our filename matching could not find the correct file anymore.

# Solution
The solution is to find the reporting graph json file by its `file_role` rather than the `file_name` as it should be more stable.

Reviewed By: aclave1

Differential Revision: D31741880

fbshipit-source-id: 40dc8831a6057ab3b38c9a900e497452664ff247