BE(tgnms): Refactor routes
Summary:
Converts all routes.js files to use the new Api class.

Changes:
* new `Api` class
* setupRoutes function
* flow-typing to some files that didn't have it
* deleted the "docker" api - it doesn't appear to be used.

## Motivation
Many of our modules execute side-effects when they are imported. This makes testing harder and can cause the app to crash when just importing a module. It makes it harder to change settings at runtime and requires us to rely on ordering our imports. While this is a general problem across the codebase, the routes.js files often need to consume settings from the env and store state in memory, making them especially prone to the pitfalls.

Requirements:
* No side-effects when importing a routes.js file
* No caching variables in the module scope

Another problem that this will help to address is sharing code/state on the backend. While it would be ideal to share state in a tool like redis to enable horizontal scaling, the reality is that the NMS uses memory state and thus needs to share code and state between modules. Currently, this is done using the nodejs [module-cache](https://nodejs.org/api/modules.html#modules_caching) - Storing state privately inside modules makes it more difficult to reset state for testing and makes sharing state harder.

## Solution
Api class - routes.js files export a class which extends the Api class. It's important to understand that no code is executed and no side-effects happen when importing this class. Only once it's instantiated and its methods are called will code be executed. The api class also makes it easy to define routes, initialize the logger for the module, and asynchronously initialize some state for the module.

Api adds a number of utilities:
* `this.initLogger(filename)` to initialize the logger, instead of require('../log')(module)
* `this.createApi()` instead of import {createApi} from '../helpers/apiHelpers';

**setupRoutes**

Registering the API routes has been moved from server.js to this setupRoutes file. It will handle initializing all the apis. There's a function called routeManager inside which handles registering the routes in a familiar way (using require) and initializes them in order.

Reviewed By: tommyhuynh

Differential Revision: D31689576

fbshipit-source-id: 7abf4abd8f19406722332e3255f5f92084e6e005