Back out "fix(tgnms): upgrade @material-table/core to v4 - fixes NetworkNodesTable group sizing"
Summary: material-table-core 4.3.6 is incompatible with our current material-ui version, causing pagination to break. When I upgraded to the latest material-ui 4.x, it fixed the issue but fbcnms/ui and fbcnms/alarms both depend on 4.11, so we included 2 versions which broke the UI even more. So i'm just backing this commit out until I can get the upgrades for fbc-js-core done.

Reviewed By: tommyhuynh

Differential Revision: D31785343

fbshipit-source-id: 4d8b82a6c5a56246ecbfa1f62b5cc95910f6e4ac