fix(tgnms): fix timeouts when running startTopologyScan
Summary: startTopologyScan was called through serviceAPIUtil which makes a request to the route in apiservice/routes.js. The default apiservice route uses a timeout of API_REQUEST_TIMEOUT which is 5s by default, but topologyScan takes much longer than this. The solution was to make a new api endpoint which uses a custom timeout of 90s. Another solution would've been to allow an arbitrary timeout param in the apiservice/routes endpoint but that felt strange to me.

Reviewed By: tommyhuynh

Differential Revision: D31770175

fbshipit-source-id: 58aa165470db415cb145def7536e308004625dad