fix(tgnms-v2) Planning Service: Create plan when file_handle is ready
Summary:
Inside planningService.js we have code that uploads input files to ANP ( uploadDraftInputFiles) and we wait for the files to be READY before we create the plan (see bug for more details: T102990217). However, the READY file_status isn't completely accurate because we still need the file handle to be non null before we launch the plan.

This diff makes us wait for file_handle to be non null before we create a plan.

Reviewed By: aclave1

Differential Revision: D31998944

fbshipit-source-id: 8ac38bf4ed35584e64f9edd0b64f67148e0c056c