fix(tgnms-v2): Switching plans resets the pending topology
Summary:
# Overview
I noticed a bug that would happen when I would open a plan and make selections (adding to my pendingTopology). Then switch to open a different plan, the NMS would crash.

This is because `pendingTopology` is a context value. When we switched plans, we never reset `pendingTopology`, thus the new plan would open with `pendingTopology` containing elements of the old plan.

# Solution
Reset `pendingTopology` when `PlanResultsView` is unmounted.

P.S. I also made some changes so that if users click the Commit Plan to Network button, if they haven't made any selections, then it defaults to the whole topology.

Reviewed By: aclave1

Differential Revision: D32001234

fbshipit-source-id: 8249c6fe70828820d244c73626eb8e5c0847a85e