fix(tgnms-v2) Planning: Copying a plan copies the input files as well.
Summary:
# Overview
When a user creates a new plan by copying an existing one, the new plan does not contain the input files. This only concerns copying a plan via CreatePlanModal component (this is used **only in the tables**). Copying a plan via the side details panel works fine.

Turns out we were never passing along the input files when we copied a plan.

# Solution
Pass in the input files to the new plan + tests

Reviewed By: aclave1

Differential Revision: D32119113

fbshipit-source-id: 519c09b8fd368568e5529736668855fc7fd56f52