feat(network_test): allow TCP link tests
Summary:
OEMs want the ability to run TCP link tests due to constraints on their hardware. This diff no longer overwrites the protocol that is supplied in the API. Instead, we set the protocol to UDP only if one isn't provided.

This behavior diverges from the NMS UI which only allows a user to select UDP iirc aclave1.

Reviewed By: silver23arrow

Differential Revision: D32153025

fbshipit-source-id: f7380cf3f10bd4fb612dcd4c9f23a2363f55f30f