feat(network_test): add traffic direction option
Summary:
By default, all network tests run simultaneous bidirectional iperf traffic on each asset in the test. Instead of running the traffic simultaneously, some OEMs would like to support running the uplink + downlink traffic streams serially.

This diff adds a new field for all network tests called `direction` and a column to the `NetworkTestParams` table called `direction`. The Alembic migration will make all current tests in the DB use the `BIDIRECTIONAL_PARALLEL` mode, as it has the same semantic meaning as the default option prior to this diff.

The `estimated_duration` method for all tests will now multiply by 2 if the mode is `BIDIRECTIONAL_SEQUENTIAL` because each asset is now processed in two goes (uplink, then downlink).

If there is a desire to ever run traffic unidirectionally (JUST uplink or JUST downlink), this diff is written in a way to support it extensibly.

Reviewed By: silver23arrow, tommyhuynh

Differential Revision: D32160840

fbshipit-source-id: 1465cfbbb3a131a1b2352b22d55673ea54a8da34