fix(network_test): set throughput ratio to 1 when supplied bitrate is 0
Summary:
One of the components in computing link health is the measured iperf throughput in relation to the expected throughput that was supplied in the initial test execution.

When passing in a bitrate of 0, the user is asking the node to disable bitrate limits and pump traffic at the maximum bandwidth. In this case, computing a throughput ratio doesn't make sense because there is no real expectation of what the throughput will be going into the test. Furthermore, the code fails in this case because it computes a throughput ratio anyways, where 0 is the denominator.

This diff changes the `iperf_tput_ratio` to be 1 when the bitrate is 0:
1. Because it doesn't make sense to compute a throughput ratio anyways
2. Because a value of 1 will short-circuit all of the health conditional checks below to always evaluate to true, essentially ignoring this component when computing health.

Reviewed By: silver23arrow

Differential Revision: D32289096

fbshipit-source-id: c8d3d4ef3b799e6aaeb8488684eda9b8d7c0daa4