fix(tgnms-v2) Default Routes: Fix start and end time inconsistencies in requests
Summary:
# Overview
The main bug reported was we were showing inconsistent info for the same day in the UI.

This was because how we calculate start and end time when the component is **first rendered** is different than how we do it when we use the **datepicker**.

You can see this by loading the component (it should automatically select today's date), then going into the date picker and selecting today's date again. Notice that the request re-fires with a different time (see test plan).

Also:
1. Cleaned up some unnecessary API calls
2. Fixed Daylight Savings Times bug where clicking on 11/5 would show 11/4 as selected on the date picker. This was because previously we were subtracting the timezone offset based on the timezone offset `TODAY`. Which would subtract an extra hour, if the day was pre-DST, making the date 11/4 23:00:00.

# Solution
This is all fixed by being more consistent with how we construct the start and end time.

Essentially before we were trying to "replicate" UTC time by doing this error prone math of subtracting/adding the timezone offset from todays time.

Now we construct the time in UTC right-off-the-bat. So that all requests now consistently end with `hrs:min:sec` as `00:00:00`. This fixes the main issue and #2.

Reviewed By: aclave1

Differential Revision: D32317415

fbshipit-source-id: 8ac04e8810c128f217a5572d42195b71e968f39d