fix(tgnms): Fix Link Utilization overlays
Summary:
Fixes the link utilization overlays for both current and historical data.

**Overlays:**
* Link Utilization (Mbps)
* Link Utilization (MCS rate)

## The problem
rx_bytes and tx_bytes are counters, but they were being interpreted as rates by the nms.

## Implementation
* Adds a new `query` function to the overlays which allows for a custom prometheus query **in-addition-to** the simple metrics. More info on overlays is below in the **Overlays** section
* the query is now based on `rate()` since rx_bytes and tx_bytes are counters

# Overlays

In TGNMS, a topology overlay takes metric/topology data and adds colors/labels to topology elements on the map. The Overlay type is the definition for the overlays gives them labels, a legend, color ranges, and tells NMS which data to use to calculate the overlay.

Metric data can be pulled for an overlay in 3 ways:
* `metrics` array, a list of prometheus metric names
* `id` - the overlay id can be used for a single-metric overlay
* `query` (**new**)- a function which returns a custom prometheus query. The query is mapped back to the overlay by appending the label  "__name__: < overlay id>" to the metrics returned by this query.

Both/either "metrics" and query can be provided. They will be merged into the "metricData" object passed to aggregate.

Reviewed By: tommyhuynh

Differential Revision: D32393738

fbshipit-source-id: 349e054e71a32a20405615996ce896abe070d26f