fix(tgnms-v2) Default Routes: Safely access routes value
Summary:
# Overview
Nathan was experiencing some errors on fresh new network where the whole app would just crash when trying to view the routes for a node. This was because the default routes service could send us back `null` and `[]` results; which we haven't accounted for.

Though according to Nathan, in theory we should get back at least one route, the default route.

# Solution
Safely access the values.

Reviewed By: aclave1

Differential Revision: D32445117

fbshipit-source-id: 37094ecbc3cd738949e1c74774fbd3b028378e49