BE(tgnms): develop.py - make project a Click choice
Summary:
Converts the --project option to a `click.Choice`. This provides us with better error messages and integrates with the --help command.

## Context:
For the `go` and `reset` commands, the --project arg is used to select which `project` to build and deploy to the test cluster.
`project` is any key in the `projects` object.

Reviewed By: tommyhuynh

Differential Revision: D32599319

fbshipit-source-id: 0a01835319729ec2dbf19fcc13567642e9257694