fix(tgnms): fix network table heights by measuring elements
Summary:
material-table's maxBodyHeight and minBodyHeight are the only way to set the table's height. Previously we calculated the heights manually by subtracting constants. This approach was brittle and needed to be updated by hand for each table.

Solution:
The solution is to measure 3 heights:
1. maxHeight: The measured height of the wrapper around MaterialTable, this is the maximum available space
2. tableHeight: The measured height of the table itself, including pagination, columns, etc
3. bodyHeight: The measured height of the body

The desired body height is then `maxHeight - (tableHeight - bodyHeight)`.

https://pxl.cl/1V73f

Caveats:
* Seems like the 100% height div wrapping all the network tables needs to have only one child, the MaterialTable. For example, the "Link Options" and "Link Table" components needed to be moved into the toolbar of the table, instead of siblings.

Reviewed By: tommyhuynh

Differential Revision: D32851318

fbshipit-source-id: 21ed03cd24d79c04cd4a5456bed0eb47aac60c58