feature(tgnms-v2) L2 Tunnels: Allow editing
Summary:
This diff allows users to edit the L2 tunnels they have created.

# Some notable features
- Users can't choose a tunnel name that already exists on the node(s). Submit button becomes disabled if invalid name is inputted.
- When editing, users can NOT change the Tunnel name or Node 1.

# Some notable missing things
As of right now, there's no ability to delete a tunnel. (I'm working on this next.)
Similarly, on a Node-to-Node, switching node2 to a different node doesn't delete the old one.

Also, converting a Node-to-Node tunnel to an External IP tunnel is tricky....if i create tunnel from `A->B`, then theres also a tunnel from `B->A`. If I convert tunnel `A->B` to `A->IP_ADDRESS`, then what happens to `B->A`? Some options:
- Don't allow switching from N2N to N2IP (easiest)
- Delete `B->A`

**Which one should we do? Preferably 1**

Reviewed By: aclave1

Differential Revision: D32778290

fbshipit-source-id: 0e5123b8f272fd30404e17036c8a8882030252ca