fix(tgnms-v2) Topology Table: Fix table height set to 0
Summary:
Previously, there were changes that allowed for a MaterialTable's height to be dynamically adjusted to the available space. This, however, resulted in a bug with the TopologyTable. The logic for dynamically changing the height did not work for tables hidden inside Accordions (specifically it wasn't able to accurately capture the available height for the table).

This diff just manually sets the height for the TopologyTables because users can't really change the size of the accordion layers anyways, so no need to complicate the autosize feature.

Reviewed By: aclave1

Differential Revision: D33794987

fbshipit-source-id: da803fd8f0d57e15ef4bb052da26a158224306e1