fix(tgnms-v2) Fix Policing Classification form for nodes
Summary:
So this feature was straight up broken. As seen in the BEFORE video below:
1. Not able to see the Simple form values (because the code never populated the value for the input fields)
2. Not able to edit in the Custom view since the field properties were wrong. The form layout itself is wrong, theres no way for users to edit each traffic class. Also the input is broken, i can't type anything in there.

# Solution
So it seems as though these forms were never designed to support multiple views for a form field (i.e. simple and custom). For one, we never saved which view the user chose, so it always defaults to Simple view.

After talking with Nathan, we decided to just get rid of the simple view (since to do it right it would likely require large architectural changes) and just have the Custom view by default.

Reviewed By: aclave1

Differential Revision: D33852969

fbshipit-source-id: 4b37d4c442d506e3b66247aee1155e609e14aca9