feat(tgnms): sites file editor
Summary:
Adds an editor for the sites file. Previously users had to download the sites-file, edit it in an external program, and reupload it. Now it can be edited directly from the NMS.

{F695985117}

## Sitesfile Editor
* drag to move sites on the map
* create and bulk copy/delete sites

# Other changes

**Manage input file**
Replaces the old browse button with an autocomplete.

{F695985593}
{F695985594}

**URLs**

Previously the PlansTable would navigate to /plan if the plan was successful, rendering the TopologyTable. Now it navigates to /plan and renders the PlanView. PlanView renders the TopologyTable if the plan is successful, SitesFileEditor if the plan is in the draft state, and redirects away from /plan otherwise.

Success
{F695986349}
Error
{F695986355}
Draft
{F695986354}

Reviewed By: tommyhuynh

Differential Revision: D31818889

fbshipit-source-id: 0f8d406e729144d79da8eb41c3abee759c83457f