fix(nms-installer) Fix Audit log returning no results for index fluentd-log-server-*
Summary:
There was an issue where the logs were not showing up for the `fluentd-log-server-*` index (on the Audit Log kibana page).

This was because there were permission issues with the `audit.log.pos` file (which tracks the position of the log read)
{F699842629}

The solution was just to modify the permissions of that file. The reason why this suddenly happened is unknown.

Reviewed By: aclave1

Differential Revision: D34122525

fbshipit-source-id: 2f37d06ce03bff997e0bf9b71038f26285c96186